+++ api/routers/admin/group/list.js
@@ -1,34 +1,23 @@
-const ObjectId = require('mongodb').ObjectID
 const Route = require('lib/router/route')
+const QueryParams = require('lib/router/query-params')
 
 const {Group, User} = require('models')
+
+const queryParams = new QueryParams()
+
+queryParams.addFilter('user', async function (filters, value) {
+  const user = await User.findOne({'uuid': value})
+
+  if (user) {
+    filters.users = { $in: [user._id] }
+  }
+})
 
 module.exports = new Route({
   method: 'get',
   path: '/',
   handler: async function (ctx) {
-    var filters = {}
-    for (var filter in ctx.request.query) {
-      if (filter === 'limit' || filter === 'start' || filter === 'sort') {
-        continue
-      }
-
-      if (filter === 'user') {
-        const user = await User.findOne({'uuid': ctx.request.query[filter]})
-
-        if (user) {
-          filters['users'] = { $nin: [ObjectId(user._id)] }
-        }
-
-        continue
-      }
-
-      if (!isNaN(parseInt(ctx.request.query[filter]))) {
-        filters[filter] = parseInt(ctx.request.query[filter])
-      } else {
-        filters[filter] = ctx.request.query[filter]
-      }
-    }
+    const filters = queryParams.toFilters(ctx.request.query)
 
     var groups = await Group.dataTables({
       limit: ctx.request.query.limit || 20,