+++ api/routers/admin/roles/list.js
@@ -1,22 +1,14 @@
 const Route = require('lib/router/route')
 const {Role} = require('models')
+const QueryParams = require('lib/router/query-params')
+
+const queryParams = new QueryParams()
 
 module.exports = new Route({
   method: 'get',
   path: '/',
   handler: async function (ctx) {
-    var filters = {}
-    for (var filter in ctx.request.query) {
-      if (filter === 'limit' || filter === 'start') {
-        continue
-      }
-
-      if (!isNaN(parseInt(ctx.request.query[filter]))) {
-        filters[filter] = parseInt(ctx.request.query[filter])
-      } else {
-        filters[filter] = ctx.request.query[filter]
-      }
-    }
+    const filters = queryParams.toFilters(ctx.request.query)
 
     var role = await Role.dataTables({
       limit: ctx.request.query.limit || 20,