+++ models/organization.js
@@ -13,6 +13,8 @@ const organizationSchema = new Schema({
   dateCreated: { type: Date, default: moment.utc },
   uuid: { type: String, default: v4 },
   isDeleted: { type: Boolean, default: false }
+}, {
+  usePushEach: true
 })
 
 organizationSchema.methods.toPublic = function () {